requirements.txt
requests
peewee
pyyaml